policywatch-backend/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
asyncpg==0.29.0
pydantic==2.8.2
feedparser==6.0.11
openai>=1.43.0
httpx==0.27.2
pdfminer.six>=20220524
python-jose[cryptography]
pypdf>=4.0.0
playwright>=1.41.0
